Fixed Order_Details table: updated quantity to DOUBLE

+++ task_2.sql
@@ -42,7 +42,7 @@ CREATE TABLE IF NOT EXISTS Order_Details (
     order_detail_id INT AUTO_INCREMENT PRIMARY KEY,
     order_id INT NOT NULL,
     book_id INT NOT NULL,
-    quantity INT NOT NULL CHECK (quantity > 0),
+    quantity DOUBLE NOT NULL CHECK (quantity > 0),
     subtotal DECIMAL(10,2) NOT NULL,
     FOREIGN KEY (order_id) REFERENCES Orders(order_id) ON DELETE CASCADE,
     FOREIGN KEY (book_id) REFERENCES Books(book_id) ON DELETE CASCADE
